Move OneHotSwitch to coreblocks.utils
